# Optimize dictFind with prefetching and branch prediction hints.