This pull request optimizes the `dictFind` function by adding software prefetching and branch prediction hints to improve cache efficiency and reduce memory latency.
It introduces 2 prefetch hints (read/write) that became no-ops in case the compiler does not support it.

 Baseline profiling with Intel VTune indicated that dictFind was significantly back-end bound, with memory latency accounting for 59.6% of clockticks, with frequent stalls from DRAM-bound operations due to cache misses during hash table lookups.

![microarch](https://github.com/user-attachments/assets/9e3cf334-ae6b-4767-b568-713a4ac24e87)


### Performance Impact

These optimizations improved the EXISTS benchmark from 924K ops/sec to 972K ops/sec (5% improvement), and we've confirmed that we've moved from overall 33% memory bound on this use-case to 31% memory bound (2% drop) while improving the CPI from 0.85 to 0.82 ).

To reproduce benchmarks:
Setup step
```
memtier_benchmark  --ratio 1:0 -n allkeys --key-pattern P:P --key-maximum 50000000 -d 10 --key-prefix "" --hide-histogram  --pipeline 100 -S /tmp/1.socket
```

baseline (06b144aa097e968f579048800cd0e5c9336ea1c9)
```
taskset -c 2-5 memtier_benchmark --key-minimum=1 --key-maximum 50000000 --key-prefix "" --command "EXISTS __key__" --pipeline 10 --command-key-pattern="R" -c 20 -t 10 --hide-histogram --test-time 120 --json-out-file baseline.json -S /tmp/1.socket --distinct-client-seed
Json file baseline.json created...
Writing results to stdout
[RUN #1] Preparing benchmark client...
[RUN #1] Launching threads now...
[RUN #1 100%, 120 secs]  0 threads:   110909270 ops,  919399 (avg:  924161) ops/sec, 29.62MB/sec (avg: 29.77MB/sec),  2.17 (avg:  2.16) msec latency

10        Threads
20        Connections per thread
120       Seconds


ALL STATS
==================================================================================================
Type         Ops/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
--------------------------------------------------------------------------------------------------
Existss    924126.70         2.16130         2.19100         2.73500         9.02300     30483.37 
Totals     924126.70         2.16130         2.19100         2.73500         9.02300     60966.75 
Json file closed.
```

comparison: (0b34249877a70b9260d932fd2fd7828cd130bddc)

```
taskset -c 2-5 memtier_benchmark --key-minimum=1 --key-maximum 50000000 --key-prefix "" --command "EXISTS __key__" --pipeline 10 --command-key-pattern="R" -c 20 -t 10 --hide-histogram --test-time 120 --json-out-file comparison.json -S /tmp/1.socket --distinct-client-seed
Json file comparison.json created...
Writing results to stdout
[RUN #1] Preparing benchmark client...
[RUN #1] Launching threads now...
[RUN #1 100%, 120 secs]  0 threads:   116747570 ops,  970742 (avg:  972811) ops/sec, 31.27MB/sec (avg: 31.34MB/sec),  2.06 (avg:  2.05) msec latency

10        Threads
20        Connections per thread
120       Seconds


ALL STATS
==================================================================================================
Type         Ops/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
--------------------------------------------------------------------------------------------------
Existss    972798.29         2.05308         2.14300         2.67100         8.57500     32088.86 
Totals     972798.29         2.05308         2.14300         2.67100         8.57500     64177.72 
Json file closed.
````