@fcostaoliveira 2024-11-08
### CE Performance Automation : step 1 of 2 (build) STARTING...

This comment was automatically generated given a benchmark was triggered.
Started building at 2024-11-08 20:29:31.805975
You can check each build/benchmark progress in grafana:
   - git hash: e0d5c42364d69ece513e880b26c25169cd96ebbf
   - git branch: filipecosta90:dict.memory.optimize
   - commit date and time: n/a
   - commit summary: n/a
   - test filters:
       - command priority lower limit: 0
       - command priority upper limit: 10000
       - test name regex: .*
       - command group regex: .*